. Included new TSPLike format for Arc Routing Problems with Time dependency based on real maps . Added parameter -format where 0 = carp instance and 1 = tsplike instance . Added parameter carp instances with oriented graphs -dir 0 = undirected graph and 1 = directed graph . Fixed bug: stop current run when params->borne >= 100.0 in function Individu::generalSplit() - the test was after the loop and could not be reached.

+++ Program/Noeud.cpp
@@ -27,7 +27,8 @@ Noeud::Noeud(bool estUnDepot, int cour,int jour, bool estPresent, Noeud * suiv ,
 	if (estUnDepot) ccour = 0 ;
 
 	// Initialization of the coutInsertion structure
-	for (int i=0 ; i < params->cli[ccour].visits.size() ; i++)
+	//for (int i=0 ; i < params->cli[ccour].visits.size() ; i++)
+	for (size_t i = 0; i < params->cli[ccour].visits.size(); i++)
 	{
 		coutInsertion.push_back(1.e30) ;
 		placeInsertion.push_back(NULL) ;
